backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.1
python-dotenv==1.0.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Gemini AI
google-generativeai==0.3.2

# Background Jobs
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
